hash160: remove the btcsuite/golangcrypto dependency.
The btcsuite/golangcrypto dependency contains a subset of the code
from golang.org/x/crypto, which is also an (indirect) dependency.

Golangcrypto is only used by this repo, all the other Decred repos
use crypto.

This commit changes the one point in main.go that uses golangcrypto
to use crypto, and deletes the hash160.go file.

It also removes golangcrypto from the glide.lock and glide.yaml files,
and adds the missing entries to the latter.

Finally, it adds the -verify option to the usage string.