Add option to spit key pair in stdout
This adds an option to just write the key pair in stdout and not write it out into a file. This is more convenient if we're calling this from a script in another runtime (e.g. Shell or Python) because we don't have to write to file and read from it just to get the address and private key.